requirements.txt
flask
redis
postgres
python-dotenv==0.20.0